requirements.txt
python-telegram-bot==20.3
requests
pydub
SpeechRecognition
gTTS
